+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tag filter component
 - Google firebase functions
 - Skeleton component
+- Stream Card component
 
 ## [0.1.0]
 ### Added